v0.4
In the GutBuster() method, a variable called force was added to hold the force that is applied to the spine rigidbody. The hardcoded value of 500 has been replaced with the force variable, which can be adjusted as needed.
A check for the time passed since the start of the script has been added in the Update() method, script will be destroyed if 5 seconds have passed after the script started.
These changes make the script more readable and user-friendly. The GutBuster() method can now be easily adjusted to apply different amounts of force to the spine rigidbody, and the time check allows the script to be destroyed after a set period of time regardless of the ragdoll's state.